+++ zarrs/examples/custom_data_type_uint4.rs
@@ -16,7 +16,7 @@ use zarrs::{
     array_subset::ArraySubset,
 };
 use zarrs_data_type::{
-    DataTypeExtension, DataTypeFillValueError, DataTypeFillValueMetadataError, DataTypePlugin,
+    DataTypeFillValueError, DataTypeFillValueMetadataError, DataTypePlugin, DataTypeTraits,
     FillValue,
 };
 use zarrs_plugin::{PluginCreateError, PluginMetadataInvalidError, ZarrVersions};
@@ -49,14 +49,14 @@ fn default_name_custom_dtype(_version: ZarrVersions) -> Cow<'static, str> {
 
 fn create_custom_dtype(metadata: &MetadataV3) -> Result<DataType, PluginCreateError> {
     if metadata.configuration_is_none_or_empty() {
-        Ok(Arc::new(CustomDataTypeUInt4))
+        Ok(Arc::new(CustomDataTypeUInt4).into())
     } else {
         Err(PluginMetadataInvalidError::new(UINT4, "codec", metadata.to_string()).into())
     }
 }
 
 /// Implement the core data type extension methods
-impl DataTypeExtension for CustomDataTypeUInt4 {
+impl DataTypeTraits for CustomDataTypeUInt4 {
     fn identifier(&self) -> &'static str {
         UINT4
     }
@@ -229,7 +229,7 @@ fn main() {
     let array = ArrayBuilder::new(
         vec![6, 1], // array shape
         vec![5, 1], // regular chunk shape
-        Arc::new(CustomDataTypeUInt4) as DataType,
+        Arc::new(CustomDataTypeUInt4),
         FillValue::new(fill_value.to_ne_bytes().to_vec()),
     )
     .array_to_array_codecs(vec![